Fix glHint parameter, type size and add more null checks (#130)

+++ source/Irrlicht/CImageLoaderPNG.cpp
@@ -40,7 +40,7 @@ void PNGAPI user_read_data_fcn(png_structp png_ptr, png_bytep data, png_size_t l
 
 	// changed by zola {
 	io::IReadFile* file=(io::IReadFile*)png_get_io_ptr(png_ptr);
-	check=(png_size_t) file->read((void*)data,(u32)length);
+	check=(png_size_t) file->read((void*)data, length);
 	// }
 
 	if (check != length)
